fix: move assets to src folder
