+++ .prettierrc
@@ -17,6 +17,7 @@
 		"^~/utils/(.*)$",
 		"^~/hooks/(.*)$",
 		"^~/lib/(.*)$",
+		"^~/assets/(.*)$",
 		"^~/styles/(.*)$",
 		"^[./]"
 	],
