+++ src/components/error/error.tsx
@@ -1,6 +1,7 @@
 import { AppRoutes } from '~/constants';
 
-import NotFoundImage from '../../../public/images/404.png';
+import NotFoundImage from '~/assets/images/404.png';
+
 import { Button } from '../button';
 import { Image } from '../image';
 import { Typography } from '../typography';
